requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-redis==5.3.0
gunicorn==20.1.0